test-asyncio-taskgroup/requirements.txt
uvicorn
fastapi==0.110.0
starlette==0.36.3
aiohttp

opentelemetry-distro
opentelemetry-instrumentation-fastapi

-e  ../../../code/sentry-python

ipdb